requirements.txt
openai
python-dotenv
numpy
sounddevice
wavio
playsound
ipywidgets
tiktoken
sentence-transformers
PyPDF2
requests
beautifulsoup4